Use the correct datatype for multistate clusters

+++ zigpy/zcl/clusters/general.py
@@ -1683,7 +1683,7 @@ class MultistateInput(Cluster):
             id=0x0051, type=t.Bool, access="r*w", mandatory=True
         )
         present_value: Final = ZCLAttributeDef(
-            id=0x0055, type=t.Single, access="r*w", mandatory=True
+            id=0x0055, type=t.uint16_t, access="r*w", mandatory=True
         )
         # 0x0057: ('priority_array', TODO.array),  # Array of 16 structures of (boolean,
         # single precision)
@@ -1716,13 +1716,13 @@ class MultistateOutput(Cluster):
             id=0x0051, type=t.Bool, access="r*w", mandatory=True
         )
         present_value: Final = ZCLAttributeDef(
-            id=0x0055, type=t.Single, access="r*w", mandatory=True
+            id=0x0055, type=t.uint16_t, access="r*w", mandatory=True
         )
         # 0x0057: ZCLAttributeDef('priority_array', type=TODO.array),  # Array of 16 structures of (boolean,
         # single precision)
         reliability: Final = ZCLAttributeDef(id=0x0067, type=t.enum8, access="r*w")
         relinquish_default: Final = ZCLAttributeDef(
-            id=0x0068, type=t.Single, access="r*w"
+            id=0x0068, type=t.uint16_t, access="r*w"
         )
         status_flags: Final = ZCLAttributeDef(
             id=0x006F, type=t.bitmap8, access="r", mandatory=True
@@ -1752,13 +1752,13 @@ class MultistateValue(Cluster):
             id=0x0051, type=t.Bool, access="r*w", mandatory=True
         )
         present_value: Final = ZCLAttributeDef(
-            id=0x0055, type=t.Single, access="r*w", mandatory=True
+            id=0x0055, type=t.uint16_t, access="r*w", mandatory=True
         )
         # 0x0057: ZCLAttributeDef('priority_array', type=TODO.array),  # Array of 16 structures of (boolean,
         # single precision)
         reliability: Final = ZCLAttributeDef(id=0x0067, type=t.enum8, access="r*w")
         relinquish_default: Final = ZCLAttributeDef(
-            id=0x0068, type=t.Single, access="r*w"
+            id=0x0068, type=t.uint16_t, access="r*w"
         )
         status_flags: Final = ZCLAttributeDef(
             id=0x006F, type=t.bitmap8, access="r", mandatory=True
